animals request prepare

+++ database/migrations/2023_05_29_105412_create_animals_table.php
@@ -41,7 +41,7 @@ return new class extends Migration
             $table->string('breed_name')->nullable();
             $table->boolean('metis')->default(0);
 
-            $table->float('weight');
+            $table->double('weight');
             $table->enum('weight_unit', array_column(WeightUnitEnum::cases(), 'name'));
 
             $table->timestamps();
